Add flexibility to URL filename discovery
When a URL is provided as a link, the code's expecting the server to give a content-disposition header to suggest a filename.  I can see why that's being done but suggest a fallback to just getting basename() of the URL if the result is blank (e.g. the server doesn't serve that optional header).